Fix eslint findings in integration tests
