--- test/integration/step_4_update.js
@@ -1,11 +1,0 @@
-const LeonardoIoT = require('../../lib/LeonardoIoT');
-
-describe('4) UPDATE', function () {
-    let client;
-
-    before(function () {
-        client = new LeonardoIoT();
-    });
-
-    it('none', function () {});
-});
