+++ test/integration/step_4_update.test.js
@@ -1,0 +1,13 @@
+const LeonardoIoT = require('../../lib/LeonardoIoT');
+
+describe('4) UPDATE', function () {
+  let client;
+
+  // eslint-disable-next-line mocha/no-hooks-for-single-case
+  before(function () {
+    // eslint-disable-next-line no-unused-vars
+    client = new LeonardoIoT();
+  });
+
+  it('none', function () { });
+});
